Add ObjectFactory.release() and releaseAll()
Methods will be used to complete the lifecycle
for when instances (or supporting objects)
will need to be called back.

releaseInstance(T instance, RuntimeContext) is
intended to release single objects, mainly
during request scoped objects for instance.

releaseAll(RuntimeContext) is intended to
release all objects currently known to the
ObjectFactory during the cycle.

It is up to each ObjectFactory to determine
the implementation.  A null passed in through
RuntimeContext can indicate that it is
outside any single context.  For example,
a releaseAll(null) could be the callback
for a singleton ObjectFactory which means
this is the last time the ObjectFactory
references are intended to be used.

git-svn-id: https://svn.apache.org/repos/asf/incubator/wink/trunk@952028 13f79535-47bb-0310-9956-ffa450edef68